#7 creating ProcessInvoiceDetail to read deatil from any kind of input
