+++ src/main/groovy/com/makingdevs/routes/ZipRoute.groovy
@@ -4,6 +4,8 @@ import com.makingdevs.routes.utils.ProcessAttachments
 import org.apache.camel.LoggingLevel
 import org.apache.camel.builder.RouteBuilder
 import org.apache.camel.dataformat.zipfile.ZipSplitter
+import org.apache.camel.Exchange
+import com.makingdevs.routes.utils.ProceessInvoiceDetail
 
 /**
  * Created by makingdevs on 3/22/17.
@@ -21,7 +23,8 @@ class ZipRoute extends RouteBuilder {
     from("direct:processZip")
       .routeId("processZip")
       .split(new ZipSplitter())
-        .streaming()
-        .to("direct:storeInLocal")
+      .streaming()
+      .process(new ProceessInvoiceDetail())
+      .to("direct:storeInLocal")
   }
 }
